+++ includes/class-wp-statistics-hits.php
@@ -101,7 +101,7 @@ class Hits extends Singleton
      */
     public function set_page_uri($page_uri)
     {
-        return isset($this->rest_hits->page_uri) ? base64_decode($this->rest_hits->page_uri) : $page_uri;
+        return isset($this->rest_hits->page_uri) ? sanitize_url(base64_decode($this->rest_hits->page_uri)) : $page_uri;
     }
 
     /**